[O11Y-199] generate column orders in file metadata
