+++ writer_test.go
@@ -82,8 +82,8 @@ var writerTests = []struct {
 		},
 		dump: `row group 0
 --------------------------------------------------------------------------------
-first_name:  BINARY ZSTD DO:4 FPO:55 SZ:90/72/0.80 VC:3 ENC:PLAIN,RLE, [more]...
-last_name:   BINARY ZSTD DO:94 FPO:143 SZ:86/68/0.79 VC:3 ENC:PLAIN,RL [more]...
+first_name:  BINARY ZSTD DO:4 FPO:55 SZ:90/72/0.80 VC:3 ENC:RLE,PLAIN, [more]...
+last_name:   BINARY ZSTD DO:94 FPO:143 SZ:86/68/0.79 VC:3 ENC:RLE,PLAI [more]...
 
     first_name TV=3 RL=0 DL=0 DS: 3 DE:PLAIN
     ----------------------------------------------------------------------------
@@ -119,8 +119,8 @@ value 3: R:0 D:0 V:Skywalker
 		},
 		dump: `row group 0
 --------------------------------------------------------------------------------
-first_name:  BINARY ZSTD DO:4 FPO:55 SZ:86/77/0.90 VC:3 ENC:RLE_DICTIO [more]...
-last_name:   BINARY ZSTD DO:90 FPO:139 SZ:82/73/0.89 VC:3 ENC:RLE_DICT [more]...
+first_name:  BINARY ZSTD DO:4 FPO:55 SZ:86/77/0.90 VC:3 ENC:RLE,RLE_DI [more]...
+last_name:   BINARY ZSTD DO:90 FPO:139 SZ:82/73/0.89 VC:3 ENC:RLE,RLE_ [more]...
 
     first_name TV=3 RL=0 DL=0 DS: 3 DE:PLAIN
     ----------------------------------------------------------------------------
@@ -176,9 +176,9 @@ value 3: R:0 D:0 V:Skywalker
 --------------------------------------------------------------------------------
 contacts:
 .name:              BINARY UNCOMPRESSED DO:0 FPO:4 SZ:75/75/1.00 VC:3  [more]...
-.phoneNumber:       BINARY SNAPPY DO:0 FPO:79 SZ:58/56/0.97 VC:3 ENC:PLAIN,RLE [more]...
-owner:              BINARY ZSTD DO:0 FPO:137 SZ:69/60/0.87 VC:2 ENC:PLAIN,RLE [more]...
-ownerPhoneNumbers:  BINARY GZIP DO:0 FPO:206 SZ:95/70/0.74 VC:3 ENC:PLAIN,RLE [more]...
+.phoneNumber:       BINARY SNAPPY DO:0 FPO:79 SZ:58/56/0.97 VC:3 ENC:RLE,PLAIN [more]...
+owner:              BINARY ZSTD DO:0 FPO:137 SZ:69/60/0.87 VC:2 ENC:RLE,PLAIN [more]...
+ownerPhoneNumbers:  BINARY GZIP DO:0 FPO:206 SZ:95/70/0.74 VC:3 ENC:RLE,PLAIN [more]...
 
     contacts.name TV=3 RL=1 DL=1
     ----------------------------------------------------------------------------
